recommended configs per gpu
