+++ scripts/llm/performance/pretrain_llama3_8b.py
@@ -12,70 +12,48 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from os.path import basename
-from typing import Optional
+from os.path import basename, splitext
 
 import nemo_run as run
-from utils import get_comm_overlap_callback_idx, hf_tokenizer, parse_cli_args, slurm_executor
+from utils import get_comm_overlap_callback_idx, hf_tokenizer, slurm_executor, set_recipe_primary_configs, get_performance_configs
+from argument_parser import parse_cli_args
 
 from nemo.collections.llm.recipes.llama3_8b import pretrain_recipe
 from nemo.collections.llm.recipes.precision.mixed_precision import bf16_with_fp8_mixed
 from nemo.lightning.pytorch.callbacks.garbage_collection import GarbageCollectionCallback
 from nemo.lightning.run.plugins import NsysPlugin, PerfEnvPlugin
 
-NUM_NODES = 1
-NUM_GPUS_PER_NODE = 8
-MICRO_BATCH_SIZE = 1
-GLOBAL_BATCH_SIZE = 128
-TP_SIZE = 1
-PP_SIZE = 1
-CP_SIZE = 2
-VP_SIZE = None
-MAX_STEPS = 100
-
-
 def llama3_8b_performance_recipe(
     compute_dtype: str,
     num_nodes: int,
     num_gpus_per_node: int,
+    max_steps: int,
     mbs: int,
     gbs: int,
     tp_size: int,
     pp_size: int,
     cp_size: int,
-    vp_size: Optional[int],
-    max_steps: int,
+    vp_size: int,
+    ep_size: int,
 ):
     """
-    llama3 8b pre-train recipe aimed at achieving best possible performance.
+    llama3 8b pre-train recipe aimed at achieving best possible performance and faster 
+    overall runtime.
 
     NOTE: Use fp8 precision training with caution. It might not give desirable results.
     """
     recipe = pretrain_recipe(performance_mode=True)
+    recipe = set_recipe_primary_configs(recipe, num_nodes, num_gpus_per_node, mbs, gbs, max_steps, tp_size, pp_size, cp_size, vp_size, ep_size)
 
     # data module configs
-    recipe.data.micro_batch_size = mbs
-    recipe.data.global_batch_size = gbs
     recipe.data.num_train_samples = max_steps * gbs * mbs  # ensure only 1 epoch for whole run
     recipe.data.tokenizer = hf_tokenizer("meta-llama/Meta-Llama-3-8B")
-
-    recipe.trainer.max_steps = max_steps
-    recipe.trainer.num_nodes = num_nodes
-    recipe.trainer.devices = num_gpus_per_node
-
-    # parallelism configs
-    recipe.trainer.strategy.tensor_model_parallel_size = tp_size
-    recipe.trainer.strategy.pipeline_model_parallel_size = pp_size
-    recipe.trainer.strategy.context_parallel_size = cp_size
-    recipe.trainer.strategy.virtual_pipeline_model_parallel_size = vp_size
-    recipe.trainer.strategy.sequence_parallel = bool(tp_size > 1)
 
     comm_overlap_callback_idx = get_comm_overlap_callback_idx(recipe.trainer.callbacks)
 
     # compute dtype configs
     if compute_dtype.lower() == "fp8":
         recipe.trainer.plugins = bf16_with_fp8_mixed()
-    recipe.trainer.plugins.grad_reduce_in_fp32 = False  # bf16 grad dtype
 
     # callback configs
     garbage_collection_callback = run.Config(
@@ -90,6 +68,7 @@ def llama3_8b_performance_recipe(
     )
     dp_size = (num_nodes * num_gpus_per_node) / (tp_size * pp_size * cp_size)
     if comm_overlap_callback_idx is not None:
+        # WARNING: If True, checkpointing (if enabled) might not work
         recipe.trainer.callbacks[comm_overlap_callback_idx].overlap_param_gather_with_optimizer_step = bool(
             dp_size > 1 and pp_size > 1 and vp_size and vp_size > 1
         )
@@ -106,22 +85,21 @@ def llama3_8b_performance_recipe(
 if __name__ == "__main__":
     args = parse_cli_args().parse_args()
 
-    exp_name = "_".join(
-        [
-            basename(__file__),
-            args.compute_dtype,
-            f"{NUM_NODES}nodes",
-            f"tp{TP_SIZE}_pp{PP_SIZE}_cp{CP_SIZE}_vp{VP_SIZE}",
-            f"{MICRO_BATCH_SIZE}mbs_{GLOBAL_BATCH_SIZE}gbs",
+    kwargs = get_performance_configs(args.gpu.lower(), "llama3", "8b", args)
+    num_nodes, mbs, gbs, tp_size, pp_size, cp_size, vp_size, ep_size = kwargs
+
+    exp_params = [
+            f"{splitext(basename(__file__))}_{args.compute_dtype}",
+            f"{num_nodes}nodes_tp{tp_size}_pp{pp_size}_cp{cp_size}_vp{vp_size}_{ep_size}_{mbs}mbs_{gbs}gbs",
         ]
-    )
+    exp_name = "_".join(exp_params)
 
     executor = slurm_executor(
         args.account,
         args.partition,
         args.log_dir,
-        NUM_NODES,
-        NUM_GPUS_PER_NODE,
+        num_nodes,
+        args.devices_per_node,
         args.time_limit,
         args.container_image,
         custom_mounts=[],
@@ -132,15 +110,16 @@ if __name__ == "__main__":
 
     recipe = llama3_8b_performance_recipe(
         args.compute_dtype,
-        NUM_NODES,
-        NUM_GPUS_PER_NODE,
-        MICRO_BATCH_SIZE,
-        GLOBAL_BATCH_SIZE,
-        TP_SIZE,
-        PP_SIZE,
-        CP_SIZE,
-        VP_SIZE,
-        MAX_STEPS,
+        num_nodes,
+        args.devices_per_node,
+        args.max_steps,
+        mbs,
+        gbs,
+        tp_size,
+        pp_size,
+        cp_size,
+        vp_size,
+        ep_size
     )
 
     if not args.tensorboard:  # tensorboard adds performance overhead.
@@ -151,17 +130,17 @@ if __name__ == "__main__":
         # following line ensures file is at- `<log_dir>/lightning_logs/tb_logs/default/<tfevents_file>`
         recipe.log.log_dir = "/nemo_run/lightning_logs"
 
-    plugins = [PerfEnvPlugin(enable_vboost=True, nccl_pp_comm_chunksize=2097152 if PP_SIZE > 1 else None)]
+    plugins = [
+        PerfEnvPlugin(
+            enable_vboost=True, 
+            nccl_pp_comm_chunksize=2097152 if pp_size > 1 else None,
+        )
+    ]
     if args.enable_profiling:
         plugins.append(NsysPlugin(start_step=5, end_step=6))
 
-    with run.Experiment(exp_name) as exp:
-        exp.add(
-            recipe,
-            executor=executor,
-            name=exp_name,
-            plugins=plugins,
-        )
+    with run.Experiment(splitext(basename(__file__))) as exp:
+        exp.add(recipe, executor=executor, name=exp_name, plugins=plugins)
 
         if not args.dryrun:
             exp.run(sequential=True, detach=True)
