+++ scripts/llm/performance/pretrain_nemotron4_15b.py
@@ -16,7 +16,8 @@ from os.path import basename
 from typing import Optional
 
 import nemo_run as run
-from utils import get_comm_overlap_callback_idx, parse_cli_args, slurm_executor
+from utils import get_comm_overlap_callback_idx, slurm_executor
+from argument_parser import parse_cli_args
 
 from nemo.collections.llm.recipes.nemotron4_15b import pretrain_recipe
 from nemo.collections.llm.recipes.precision.mixed_precision import bf16_with_fp8_mixed
