+++ scripts/llm/performance/recommended_model_configs/model_configs_b200.csv
@@ -1,0 +1,27 @@
+task,model,size,system,dtype,num_gpus,seq_len,tp_size,pp_size,cp_size,dp_size,ep_size,vp_size,mbs,gbs
+lora,llama3,70b,b200,fp8,8.0,4096.0,1.0,4.0,1.0,2.0,,20.0,1.0,32.0
+lora,llama3,70b,b200,bf16,8.0,4096.0,1.0,4.0,1.0,2.0,,20.0,1.0,32.0
+lora,llama3,8b,b200,bf16,8.0,4096.0,1.0,1.0,1.0,8.0,,,1.0,32.0
+lora,llama3,8b,b200,fp8,8.0,4096.0,1.0,1.0,1.0,8.0,,,1.0,32.0
+lora,llama31,405b,b200,fp8,24.0,2048.0,4.0,6.0,1.0,1.0,,7.0,1.0,24.0
+lora,llama31,405b,b200,bf16,24.0,4096.0,4.0,6.0,1.0,1.0,,7.0,1.0,24.0
+pre_train,gpt3,175b,b200,bf16,128.0,2048.0,4.0,4.0,1.0,8.0,,12.0,2.0,256.0
+pre_train,gpt3,175b,b200,fp8,128.0,2048.0,4.0,4.0,1.0,8.0,,12.0,2.0,256.0
+pre_train,llama3,8b,b200,bf16,8.0,8192.0,1.0,1.0,1.0,8.0,1.0,1.0,2.0,128.0
+pre_train,llama3,8b,b200,fp8,8.0,8192.0,1.0,1.0,1.0,8.0,1.0,1.0,2.0,128.0
+pre_train,llama3,70b,b200,fp8,64.0,8192.0,2.0,4.0,2.0,4.0,1.0,5.0,1.0,128.0
+pre_train,llama3,70b,b200,bf16,64.0,8192.0,2.0,4.0,2.0,4.0,1.0,5.0,1.0,128.0
+pre_train,llama31,405b,b200,bf16,144.0,8192.0,4.0,9.0,2.0,2.0,1.0,7.0,1.0,36.0
+pre_train,llama31,405b,b200,fp8,144.0,8192.0,4.0,9.0,2.0,2.0,1.0,7.0,1.0,36.0
+pre_train,mixtral,8x7b,b200,bf16,64.0,4096.0,1.0,1.0,1.0,64.0,8.0,,2.0,256.0
+pre_train,mixtral,8x7b,b200,fp8,64.0,4096.0,1.0,1.0,1.0,64.0,8.0,,2.0,256.0
+pre_train,mixtral,8x22b,b200,bf16,256.0,65536.0,2.0,4.0,8.0,4.0,8.0,14.0,1.0,64.0
+pre_train,mixtral,8x22b,b200,fp8,256.0,65536.0,2.0,4.0,8.0,4.0,8.0,14.0,1.0,64.0
+pre_train,nemotron4,15b,b200,bf16,64.0,4096.0,2.0,1.0,1.0,32.0,,,2.0,256.0
+pre_train,nemotron4,340b,b200,bf16,128.0,4096.0,8.0,4.0,1.0,4.0,,12.0,1.0,32.0
+pre_train,nemotron4,340b,b200,fp8,128.0,4096.0,8.0,4.0,1.0,4.0,,12.0,1.0,32.0
+pre_train,nemotron4,15b,b200,fp8,64.0,4096.0,2.0,1.0,1.0,32.0,,,2.0,256.0
+sft,llama3,8b,b200,fp8,8.0,4096.0,1.0,1.0,1.0,8.0,,,1.0,32.0
+sft,llama3,70b,b200,bf16,32.0,4096.0,2.0,4.0,1.0,4.0,,5.0,1.0,32.0
+sft,llama3,70b,b200,fp8,32.0,4096.0,2.0,4.0,1.0,4.0,,5.0,1.0,32.0
+sft,llama3,8b,b200,bf16,8.0,4096.0,1.0,1.0,1.0,8.0,,,1.0,32.0
